data/h3_raster_importer/requirements.txt
numpy==1.21.1
pandas==1.3.1
rasterio==1.2.6
h3ronpy==0.12.0
psycopg2-binary==2.9.1
docopt==0.6.0
